+++ Apartment_Admin/src/main/java/com/apartment/repo/ComplaintsRepo.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.apartment.models.Complaints;
 
 @Repository
-public interface ComplaintsRepo extends JpaRepository <Complaints,Integer>{
+public interface ComplaintsRepo extends JpaRepository <Complaints,Long>{
 
 }
